call to restaurant added to CategoryActivity as of now Can be added to other activities as well..

+++ app/src/main/java/in/co/tsystem/mango/mangoGlobals.java
@@ -12,7 +12,7 @@ public class mangoGlobals extends Application {
     public static String cname;
     public static String cval;
     public static Integer cookie_set = 0;
-    public static Integer total_cart_price = 0;
+    public static Double total_cart_price = 0.0;
     public static Integer no_of_items_in_cart = 0;
     private static mangoGlobals singleInstance = null;
     public static mangoGlobals getInstance() {
